# Optimize `groupby::scan`