@vuule 2022-01-07
Is there a benchmark that shows the perf benefit?