@devavret 2022-01-07
I think the original plan was better to short circuit in the sort_helper. It is also used in rolling. Otherwise, we'd have to fix it again for rolling.

Would it be possible or would it require a bigger refactor?