@PointKernel 2022-01-07
> I think the original plan was better to short circuit in the sort_helper. It is also used in rolling. Otherwise, we'd have to fix it again for rolling.
> 
> Would it be possible or would it require a bigger refactor?

@devavret It's definitely possible and actually involves fewer code changes than the current refactor. The question is really whether we want to materialize `sorted_values`/`grouped_values` or not. short circuit in `sort_helper` will materialize/memorize those two while the current refactor won't.