@devavret 2022-01-07
> short circuit in `sort_helper` will materialize/memorize those two while the current refactor won't (thus less memory footprint).

You can memoize both the view and the owning column. Anyway I now think your method would be cleaner