@PointKernel 2022-01-07
@vuule I've just added the benchmark results. This PR brings about 60% speedups for presorted scans.